# service/src/main/resources/api/openapi.yaml
---
openapi: 3.0.3
info:
  title: Quarkus and svelte demo application
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
security:
- basicAuth: []
tags:
- name: Demo
  description: Quarkus and svelte demo application
paths:
  /api/auth/refresh:
    get:
      tags:
      - Auth Resource
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        "403":
          description: Not Allowed
        "401":
          description: Not Authorized
      security:
      - SecurityScheme: []
  /api/auth/signin:
    post:
      tags:
      - Auth Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
  /hello/permit-all:
    get:
      tags:
      - Greeting Resource
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hello'
  /hello/roles-allowed:
    get:
      tags:
      - Greeting Resource
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hello'
        "403":
          description: Not Allowed
        "401":
          description: Not Authorized
      security:
      - SecurityScheme: []
components:
  schemas:
    AuthResult:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
    Hello:
      type: object
      properties:
        value:
          type: string
    Login:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT